bugfixes and simplifications removed nasty bug in signal() unfied parser and server_parser

+++ initat/host_monitoring/modules/process_monitor_mod.py
@@ -192,7 +192,7 @@ class _general(hm_classes.hm_module):
 class procstat_command(hm_classes.hm_command):
     def __init__(self, name):
         hm_classes.hm_command.__init__(self, name, server_arguments=True, positional_arguments=True)
-        self.server_parser.add_argument("-f", dest="filter", action="store_true", default=False)
+        self.parser.add_argument("-f", dest="filter", action="store_true", default=False)
         self.parser.add_argument("-w", dest="warn", type=int, default=0)
         self.parser.add_argument("-c", dest="crit", type=int, default=0)
         self.parser.add_argument("-Z", dest="zombie", default=False, action="store_true", help="ignore zombie processes")
@@ -462,8 +462,8 @@ class ipckill_command(hm_classes.hm_command):
     def __init__(self, name):
         hm_classes.hm_command.__init__(self, name, positional_arguments=True)
         self.server_arguments = True
-        self.server_parser.add_argument("--min-uid", dest="min_uid", type=int, default=0)
-        self.server_parser.add_argument("--max-uid", dest="max_uid", type=int, default=65535)
+        self.parser.add_argument("--min-uid", dest="min_uid", type=int, default=0)
+        self.parser.add_argument("--max-uid", dest="max_uid", type=int, default=65535)
     def __call__(self, srv_com, cur_ns):
         sig_str = "remove all all shm/msg/sem objects for uid %d:%d" % (
             cur_ns.min_uid,
@@ -525,10 +525,10 @@ class signal_command(hm_classes.hm_command):
     def __init__(self, name):
         hm_classes.hm_command.__init__(self, name, positional_arguments=True)
         self.server_arguments = True
-        self.server_parser.add_argument("--signal", dest="signal", type=int, default=15)
-        self.server_parser.add_argument("--min-uid", dest="min_uid", type=int, default=0)
-        self.server_parser.add_argument("--max-uid", dest="max_uid", type=int, default=65535)
-        self.server_parser.add_argument("--exclude", dest="exclude", type=str, default="")
+        self.parser.add_argument("--signal", dest="signal", type=int, default=15)
+        self.parser.add_argument("--min-uid", dest="min_uid", type=int, default=0)
+        self.parser.add_argument("--max-uid", dest="max_uid", type=int, default=65535)
+        self.parser.add_argument("--exclude", dest="exclude", type=str, default="")
         self.__signal_dict = dict([(getattr(signal, name), name) for name in dir(signal) if name.startswith("SIG") and not name.startswith("SIG_")])
     def get_signal_string(self, cur_sig):
         return self.__signal_dict.get(cur_sig, "#%d" % (cur_sig))
@@ -552,32 +552,35 @@ class signal_command(hm_classes.hm_command):
                     # do not take it
                     return 0
         # check arguments
-        exclude_list = cur_ns.exclude.split(",")
-        include_list = cur_ns.arguments
-        sig_str = "signal %d[%s] (uid %d:%d), exclude_list is %s, include_list is %s" % (
-            cur_ns.signal,
-            self.get_signal_string(cur_ns.signal),
-            cur_ns.min_uid,
-            cur_ns.max_uid,
-            ", ".join(exclude_list) or "<empty>",
-            ", ".join(include_list) or "<empty>"
-        )
-        self.log(sig_str)
+        exclude_list = [_entry for _entry in cur_ns.exclude.split(",") if _entry.strip()]
+        include_list = [_entry for _entry in cur_ns.arguments if _entry.strip()]
         srv_com["signal_list"] = []
-        pid_list = find_pids(process_tools.build_ps_tree(process_tools.get_proc_list()), priv_check)
-        for struct in pid_list:
-            try:
-                os.kill(struct["pid"], cur_ns.signal)
-            except:
-                info_str, is_error = (process_tools.get_except_info(), True)
-            else:
-                info_str, is_error = ("sent %d to %d" % (cur_ns.signal, struct["pid"]), False)
-            self.log("%d: %s" % (struct["pid"], info_str), logging_tools.LOG_LEVEL_ERROR if is_error else logging_tools.LOG_LEVEL_OK)
-            srv_com["signal_list"].append(srv_com.builder("signal", struct["name"],
-                                                          error="1" if is_error else "0",
-                                                          result=info_str,
-                                                          cmdline=" ".join(struct["cmdline"])))
-        srv_com["signal_list"].attrib.update({"signal" : "%d" % (cur_ns.signal)})
+        if not include_list and not exclude_list:
+            self.log("refuse to operate without include or exclude list", logging_tools.LOG_LEVEL_ERROR)
+        else:
+            sig_str = "signal %d[%s] (uid %d:%d), exclude_list is %s, include_list is %s" % (
+                cur_ns.signal,
+                self.get_signal_string(cur_ns.signal),
+                cur_ns.min_uid,
+                cur_ns.max_uid,
+                ", ".join(exclude_list) or "<empty>",
+                ", ".join(include_list) or "<empty>"
+            )
+            self.log(sig_str)
+            pid_list = find_pids(process_tools.build_ps_tree(process_tools.get_proc_list()), priv_check)
+            for struct in pid_list:
+                try:
+                    os.kill(struct["pid"], cur_ns.signal)
+                except:
+                    info_str, is_error = (process_tools.get_except_info(), True)
+                else:
+                    info_str, is_error = ("sent %d to %d" % (cur_ns.signal, struct["pid"]), False)
+                self.log("%d: %s" % (struct["pid"], info_str), logging_tools.LOG_LEVEL_ERROR if is_error else logging_tools.LOG_LEVEL_OK)
+                srv_com["signal_list"].append(srv_com.builder("signal", struct["name"],
+                                                              error="1" if is_error else "0",
+                                                              result=info_str,
+                                                              cmdline=" ".join(struct["cmdline"])))
+        srv_com["signal_list"].attrib.update({"signal" : "{:d}".format(cur_ns.signal)})
     def interpret(self, srv_com, cur_ns):
         ok_list, error_list = (srv_com.xpath(".//ns:signal[@error='0']/text()", smart_strings=False),
                                srv_com.xpath(".//ns:signal[@error='1']/text()", smart_strings=False))
@@ -589,24 +592,24 @@ class signal_command(hm_classes.hm_command):
             " (%s)" % (logging_tools.get_plural("problem", len(error_list))) if error_list else "")
 
 def find_pids(ptree, check):
-    def search(dict, add, start):
+    def search(_dict, add, start):
         # external check.
         # if 1 is returned, all subsequent process are added
         # if 0 is returned, the actual add-value is used
         # if -1 is returned, the add value is set to zero and all subsequent checks are disabled
-        new_add = check(start, dict[start])
+        new_add = check(start, _dict[start])
         if new_add == -1:
             add = 0
         elif new_add == 1:
             add = 1
         if add:
-            r_list, add = ([dict[start]], 1)
+            r_list, add = ([_dict[start]], 1)
         else:
             r_list = []
-        if dict[start]["childs"] and new_add >= 0:
-            p_list = dict[start]["childs"].keys()
+        if _dict[start]["childs"] and new_add >= 0:
+            p_list = _dict[start]["childs"].keys()
             for pid in p_list:
-                r_list.extend(search(dict[start]["childs"], add, pid))
+                r_list.extend(search(_dict[start]["childs"], add, pid))
         return r_list
     return search(ptree, 0, ptree.keys()[0])
